+++ src/main/java/com/iwm/backend/api/models/WeeklyScheduleEM.java
@@ -1,0 +1,19 @@
+package com.iwm.backend.api.models;
+
+import jakarta.persistence.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+public class WeeklyScheduleEM {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @OneToMany(mappedBy = "schedule", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<ShiftEM> shifts = new ArrayList<>();
+}
+
+
